backend/requirements.txt
playwright>=1.40.0
openai>=1.0.0 
flask
flask_cors
flask_jwt_extended
sqlalchemy
psycopg2-binary
python-dotenv
supabase
requests
flask_sqlalchemy
browser-use
flask-migrate
APScheduler>=3.10.0